requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==2.0
django-embed-video==1.3
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
